refactor : rankingSection 로직과 UI분리
